Support custom auth (#16)
* add config option for custom auth

* add feature flag for local-auth in Filigree

* Add custom auth test app

This doesn't build yet since the templates are still trying to generate
endpoints that use the local auth code

* Put auth provider choices into template context

* filigree crate builds with local auth disabled

* add template conditions to a lot of code that only applies for local auth

* updates halfway there

* Start converting model definitions to allow string IDs

* more work

* create schemas if needed. Use auth ID types as appropriate

* Use templated auth ID type and and schema in auth SQL templates

* more

* Handle schema creation

* fix indexes not in primary schema

* builds with custom auth and string ids

* a few more fixes

* re-add migrate_on_start to template context

+++ filigree-cli/sql/user_info.down.sql
@@ -1,15 +1,15 @@
-DROP TABLE user_invites;
+{% if auth.builtin %} DROP TABLE {{auth_schema}}.user_invites;
 
-DROP TABLE oauth_authorization_sessions;
+DROP TABLE {{auth_schema}}.oauth_authorization_sessions;
 
-DROP TABLE oauth_logins;
+DROP TABLE {{auth_schema}}.oauth_logins;
 
-DROP TABLE email_logins;
+DROP TABLE {{auth_schema}}.email_logins;
 
-DROP TABLE api_keys;
+DROP TABLE {{auth_schema}}.api_keys;
 
-DROP TABLE organization_members;
+DROP TABLE {{auth_schema}}.organization_members;
 
-DROP TABLE user_sessions;
+DROP TABLE {{auth_schema}}.user_sessions;
 
-DROP TABLE user_roles;
+{% endif %} DROP TABLE {{auth_schema}}.user_roles;
